requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
jinja2==3.1.4
python-multipart==0.0.12
python-dotenv==1.0.1
groq==0.11.0
spacy==3.7.6
requests==2.32.3
fastapi-users[sqlalchemy]
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
asyncpg
psycopg2-binary
aiomysql